Seperated Backend to its own folder, and added path pointers between files so they can be visible to each other
